+++ app/src/js/actions/index.js
@@ -2,7 +2,6 @@
 import compareVersions from 'compare-versions';
 import { get as getProperty } from 'object-path';
 import requestPromise from 'request-promise';
-import { CMR } from '@cumulus/cmrjs';
 import get from 'lodash/get';
 import isEmpty from 'lodash/isEmpty';
 import cloneDeep from 'lodash/cloneDeep';
@@ -29,13 +28,6 @@ const {
   defaultPageLimit,
   minCompatibleApiVersion
 } = _config;
-
-/**
- * match MMT to CMR environment.
- * @param {string} env - cmr environment defaults to 'SIT'
- * @returns {string} correct hostname for mmt environment
- */
-const hostId = (env = 'SIT') => getProperty({ OPS: '', SIT: 'sit', UAT: 'uat' }, env, 'sit');
 
 export const refreshAccessToken = (token) => (dispatch) => {
   const start = new Date();
@@ -136,11 +128,7 @@ export const listCollections = (options = {}) => {
         method: 'GET',
         id: null,
         url: new URL(urlPath, root).href,
-        qs: { limit: defaultPageLimit, ...queryOptions, ...timeFilters }
-      }
-    }).then(() => {
-      if (getMMT) {
-        dispatch(getMMTLinks());
+        qs: { limit: defaultPageLimit, ...queryOptions, ...timeFilters, getMMT }
       }
     });
   };
@@ -190,74 +178,6 @@ export const getCumulusInstanceMetadata = () => ({
     path: 'instanceMeta'
   }
 });
-
-/**
- * Iterates over each collection in the application collections state
- * dispatching the action to add the MMT link to the state.
- *
- * @returns {function} anonymous redux-thunk function
- */
-export const getMMTLinks = () => (dispatch, getState) => {
-  const { data } = getState().collections.list;
-  const doDispatch = ({ name, version }) => (url) => dispatch({
-    type: types.ADD_MMTLINK,
-    data: { name, version, url }
-  });
-
-  data.forEach((collection) => getMMTLinkFromCmr(collection, getState)
-    .then(doDispatch(collection))
-    .catch((error) => console.error(error)));
-};
-
-/**
- * Returns a Promise for the Metadata Management Toolkit (MMT) URL string for
- * the specified collection, or an empty promise if the collection is not found
- * in the CMR.
- *
- * @param {Object} collection - application collections item
- * @param {function} getState - redux function to access app state
- * @returns {Promise<string>} - Promise for a Metadata Management Toolkit (MMT)
- *    Link (URL string) to the input collection, or undefined if it isn't found
- */
-export const getMMTLinkFromCmr = (collection, getState) => {
-  const {
-    cumulusInstance: { cmrProvider, cmrEnvironment }, mmtLinks
-  } = getState();
-
-  if (!cmrProvider || !cmrEnvironment) {
-    return Promise.reject(
-      new Error('Missing Cumulus Instance Metadata in state.' +
-                ' Make sure a call to getCumulusInstanceMetadata is dispatched.')
-    );
-  }
-
-  if (getCollectionId(collection) in mmtLinks) {
-    return Promise.resolve(mmtLinks[getCollectionId(collection)]);
-  }
-
-  return new CMR(cmrProvider).searchCollections(
-    {
-      short_name: collection.name,
-      version: collection.version
-    }
-  )
-    .then(([result]) => result && result.id && buildMMTLink(result.id, cmrEnvironment));
-};
-
-/**
- * Returns the MMT URL string for collection based on conceptId and Cumulus
- * environment.
- *
- * @param {string} conceptId - CMR's concept id
- * @param {string} cmrEnv - Cumulus instance operating environ UAT/SIT/PROD
- * @returns {string} MMT URL string to edit the collection at conceptId
- */
-export const buildMMTLink = (conceptId, cmrEnv) => {
-  const url = ['mmt', hostId(cmrEnv), 'earthdata.nasa.gov']
-    .filter((value) => value)
-    .join('.');
-  return `https://${url}/collections/${conceptId}`;
-};
 
 export const getGranule = (granuleId) => ({
   [CALL_API]: {