+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - **CUMULUS-2063**
   - Updates the dashboard to use alpha version `@cumulus/api@3.0.1-alpha.2` for testing.
+  - Code changes to allow for private CMR collections to have links to the MMT.
 
 ## [v2.0.0]